+++ src/Motors/ServoMotor.cpp
@@ -5,22 +5,22 @@
 //* **************************** SERVO MOTOR *******************************
 //* ************************************************************************
 
-ServoMotor::ServoMotor(int pin) : servoPin(pin), currentAngle(0) {}
+ServoMotor::ServoMotor(int pin) : servoPin(pin), currentAngle(0.0f) {}
 
-void ServoMotor::init(int initialAngle) {
+void ServoMotor::init(float initialAngle) {
     servo.attach(servoPin);
     setAngle(initialAngle);
     Serial.println("Servo Initialized at: " + String(initialAngle) + " degrees");
 }
 
-void ServoMotor::setAngle(int angle) {
-    if (angle < 0) angle = 0;
-    if (angle > 180) angle = 180;
-    servo.write(angle);
+void ServoMotor::setAngle(float angle) {
+    if (angle < 0.0f) angle = 0.0f;
+    if (angle > 180.0f) angle = 180.0f;
+    servo.write((int)angle); // ESP32Servo expects int
     currentAngle = angle;
     //Serial.println("Servo moved to: " + String(angle) + " degrees"); // Avoid serial print during potential motor movement
 }
 
-int ServoMotor::getCurrentAngle() {
+float ServoMotor::getCurrentAngle() {
     return currentAngle;
 }
